pyproject.toml
[project]
name = "clinical-trial-screening-openenv"
version = "0.1.0"
description = "OpenEnv clinical trial patient screening environment"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "openenv-core>=0.2.3",
    "pydantic>=2.12.0",
    "uvicorn>=0.24.0",
]

[project.scripts]
server = "server.app:main"
